api-backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.1
google-generativeai==0.8.5
python-dotenv==1.0.1
pydantic==2.10.4 